+++ tests/src/InputHiddenSlotTest.svelte
@@ -7,6 +7,9 @@
 
 <Select bind:items name="test" bind:value>
     {#snippet inputHiddenSnippet(value)}
-        <input type="hidden" name="test" value={value && !Array.isArray(value) ? value.value : ''} />
+        <input
+            type="hidden"
+            name="test"
+            value={value && typeof value !== 'string' && !Array.isArray(value) ? value.value : ''} />
     {/snippet}
 </Select>